requirements.txt
duckduckgo_search==5.3.0
lxml==5.4.0
trafilatura==1.7.0
htmldate==1.9.3
openai>=1.40.0
google-generativeai>=0.7.2
pandas>=2.2.2
gspread>=6.1.2
google-auth>=2.32.0
vaderSentiment>=3.3.2
streamlit>=1.37.1
tqdm>=4.66.4
requests>=2.32.3
python-dotenv>=1.0.1
types-pytz

